User size_t counter in conting_sort

+++ src/graphlab/util/generics/counting_sort.hpp
@@ -42,7 +42,7 @@ namespace graphlab {
       if(value_vec.size() == 0) return;
 
       valuetype maxval = *std::max_element(value_vec.begin(), value_vec.end());
-      std::vector< atomic<int> > counter_array(maxval+1);
+      std::vector< atomic<size_t> > counter_array(maxval+1);
       permute_index.resize(value_vec.size(), 0);
       permute_index.assign(value_vec.size(), 0);
 #ifdef _OPENMP
